Simplify MekanismExternalStorageStrategy
